+++ NEWS.md
@@ -1,10 +1,16 @@
 # tsibble 0.3.0.9000
 
+## Breaking changes
+
+* Renamed `x` to `.x` in windowed functions, `slide()`, `tile()` and `stretch()`, to be consistent with purrr functions.
+* Renamed `x` to `...` in windowed functions, `slider()`, `tiler()` and `stretcher()`, to take multiple inputs.
+
 ## New features
 
 * New `holiday_aus()` for Australian national and state-based public holiday.
 * Defined `diff()` for year-week, year-month, and year-quarter.
 * `yearweek()`, `yearmonth()`, `yearquarter()` supported for character.
+* `slider()`, `tiler()`, `stretcher()` supported multiple inputs now.
 
 ## Improvements